+++ contracts/deploy/l1/01_DNSTLDResolver.ts
@@ -1,6 +1,6 @@
 import { artifacts, execute } from "@rocketh";
 import { zeroAddress } from "viem";
-import { dnsEncodeName } from "../../test/integration/utils/utils.js";
+import { dnsEncodeName } from "../../test/utils/utils.js";
 import { MAX_EXPIRY } from "../constants.js";
 
 async function fetchPublicSuffixes() {
