+++ contracts/test/integration/l1/dns/mainnet.ts
@@ -1,4 +1,7 @@
-import { COIN_TYPE_ETH, type KnownProfile } from "../../utils/resolutions.js";
+import {
+  COIN_TYPE_ETH,
+  type KnownProfile,
+} from "../../../utils/resolutions.js";
 
 export const KNOWN_DNS: KnownProfile[] = [
   {
