+++ contracts/test/integration/l1/dns/rr.ts
@@ -1,5 +1,5 @@
 import { type ByteArray, concat, stringToBytes, toHex } from "viem";
-import { packetToBytes } from "../../utils/utils.js";
+import { packetToBytes } from "../../../utils/utils.js";
 
 export type RR = {
   name: ByteArray;
